+++ CHANGELOG.md
@@ -38,7 +38,9 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
-* nothing
+### Bug Fixes
+
+* Allow marker funds to be used via feegrant again [#2110](https://github.com/provenance-io/provenance/issues/2110).
 
 ---